\name{linnerud}
\encoding{latin1}
\alias{linnerud}

\docType{data}

\title{Linnerud Dataset}

\description{
Three physiological and three exercise variables are measured on twenty middle-aged 
men in a fitness club.
}

\usage{data(linnerud)}

\format{
\itemize{
	A list containing the following components: 
	    \item \code{exercise} data frame with 20 observations on 3 exercise variables.
	    \item \code{physiological} data frame with 20 observations on 3 physiological variables. 
	}
}

\source{
Tenenhaus, M. (1998), Table 1, page 15.
}

\references{
Tenenhaus, M. (1998). \emph{La r\'egression PLS: th\'eorie et pratique}. Paris: Editions Technic.
}

\keyword{datasets}
