\name{rCCAsPLS-package}
\encoding{latin1}
\alias{rCCAsPLS-package}
\alias{rCCAsPLS}
\docType{package}
\title{
Regularized Canonical Correlation Analysis and Sparse Partial Least Squares Regression
}
\description{
The package provides a set of functions that extend the
\code{CCA} package with new numerical and graphical outputs and performe
Sparse Partial Least Squares.
}
\details{
\tabular{ll}{
Package: \tab rCCAsPLS\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-00-00\cr
License: \tab None\cr
}

}
\author{
Sébastien Déjean, Ignacio González and Kim-Anh Lê Cao.

Maintainer: Sébastien Déjean <sebastien.dejean@math.ups-tlse.fr>
}
\references{
www.lsp.ups-tlse.fr/CCA
}

\keyword{package}


