\name{scatterutil}
\alias{scatterutil.base}
\alias{scatterutil.eti}
\alias{scatterutil.grid}
\title{Graphical utility functions from ade4}
\description{
These are utilities used in graphical functions.
}

\details{
  The functions scatter use some utilities functions :
  \describe{
  \item{scatterutil.base}{defines the layer of the plot for all
    scatters}
  \item{scatterutil.eti}{puts labels centred on the points}
  \item{scatterutil.grid}{plots a grid and adds a legend}
  }
}
\seealso{
\code{\link{s.match}}
}
\author{Daniel Chessel, Stephane Dray \email{dray@biomserv.univ-lyon1.fr}}


\keyword{multivariate}
\keyword{hplot}
